refactor behavior
