+++ src/collective/judgment/browser/configure.zcml
@@ -28,7 +28,7 @@
       />
 
   <browser:viewlet
-      name="collective.judgment.voting"
+      name="collective.judgment.evaluation"
       for="collective.judgment.interfaces.IEvaluable"
       manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
       layer="..interfaces.ICollectiveJudgmentLayer"
