+++ src/collective/judgment/interfaces.py
@@ -24,48 +24,30 @@ class IEvaluable(Interface):
 # behaviors interface. When doing IEvaluation(object), you receive an adapter
 class IEvaluation(model.Schema):
     directives.omitted('evaluations')
-    directives.omitted('voted')
     if api.env.debug_mode():
         directives.no_omit(IEditForm, 'evaluations')
-        directives.no_omit(IEditForm, 'voted')
 
     fieldset(
         'debug',
         label=u'debug',
-        fields=('evaluations', 'voted')
+        fields=('evaluations', )
     )
 
     evaluations = schema.Dict(
-        title=u'Vote info',
-        key_type=schema.TextLine(title=u'voted number'),
-        value_type=schema.Int(title=u'Voted so often'),
+        title=u'Evaluation info',
+        key_type=schema.TextLine(title=u'User ID'),
+        value_type=schema.TextLine(title=u'Evaluation'),
         required=False)
 
-    voted = schema.List(
-        title=u'Vote hashes',
-        value_type=schema.TextLine(),
-        required=False)
-
-    def vote(request):
+    def evaluate(evaluation, userid):
         """
-        Store the vote information, store the request hash to ensure
+        Store the evaluation information, store the user id to ensure
         that the user does not vote twice
         """
 
-    def average_vote():
+    def already_evaluated(userid):
         """
-        Return the average voting for an item
-        """
-
-    def has_evaluations():
-        """
-        Return whether anybody ever voted for this item
-        """
-
-    def already_voted(request):
-        """
-        Return the information wether a person already voted.
-        This is not very high level and can be tricked out easily
+        Return the information wether a person already evaluated.
         """
 
     def clear():
